Add new paper: Embodied-Reasoner
Hello, thank you for maintaining this awesome repository!I’d like to recommend Embodied Reasoner, a model that extends O1-style reasoning to interactive embodied tasks.It is capable of processing lengthy, image-action interleaved contexts and producing consistent reasoning. Moreover, it generates diverse modes of thought, such as analysis, planning, and reflection, that dynamically adapt to different stages of the interaction.
🧠 Paper: https://arxiv.org/abs/2503.21696
📂 GitHub: https://github.com/zwq2018/embodied_reasoner
📊 Dataset: https://huggingface.co/datasets/zwq2018/embodied_reasoner
🌐 Website: https://embodied-reasoner.github.io/ 
Thanks in advance for considering this contribution.